requirements.txt
bertopic==0.11.0
emoji==1.7.0
pandas==1.4.3
Pillow==9.2.0
scikit_learn==1.1.2
spacy==3.4.0
streamlit==1.12.0
transformers==4.21.0
